使用ResponseEntity 包装返回值

+++ cse-account-service/src/main/java/com/huawei/account/cseaccountservice/controller/AccountController.java
@@ -1,10 +1,14 @@
 package com.huawei.account.cseaccountservice.controller;
 
+import java.util.ArrayList;
 import java.util.List;
 
+import org.apache.commons.lang.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,7 +17,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.huawei.account.cseaccountservice.entity.TAccount;
 import com.huawei.account.cseaccountservice.service.AccountService;
@@ -26,43 +29,56 @@ public class AccountController {
   @Autowired
   private AccountService accountService;
 
-  @ResponseBody
   @GetMapping("/getAllAccount")
-  public List<TAccount> getAllAccount() {
+  public ResponseEntity<List<TAccount>> getAllAccount() {
     List<TAccount> allAccounts = accountService.getAllAccounts();
     if (allAccounts == null || allAccounts.size() == 0) {
       logger.warn("accounts list is empty");
+      return ResponseEntity.ok(new ArrayList<>());
     }
-    return allAccounts;
+    return ResponseEntity.ok(allAccounts);
   }
 
-  @ResponseBody
   @PostMapping(value = "/create")
-  public String createAccount(@RequestBody TAccount account) {
-    return accountService.insertAccount(account);
+  public ResponseEntity<String> createAccount(@RequestBody TAccount account) {
+    return ResponseEntity.ok(accountService.insertAccount(account));
   }
 
-  @ResponseBody
   @DeleteMapping(value = "/delete/{id}")
-  public void createAccount(@PathVariable(value = "id") String id) {
+  public ResponseEntity<String> deleteAccount(@PathVariable(value = "id") String id) {
+    if (StringUtils.isBlank(id)) {
+      return ResponseEntity.badRequest().body(" id 不能为 空");
+    }
     accountService.deleteAccountById(id);
+    return ResponseEntity.status(HttpStatus.NO_CONTENT).body(String.format("删除 %s成功", id));
   }
 
-  @ResponseBody
   @GetMapping(value = "/search/{id}")
-  public TAccount searchAccount(@PathVariable(value = "id") String id) {
-    return accountService.findAccountById(id);
+  public ResponseEntity<TAccount> searchAccount(@PathVariable(value = "id") String id) {
+    if (StringUtils.isBlank(id)) {
+      return ResponseEntity.badRequest().build();
+    }
+    return ResponseEntity.ok(accountService.findAccountById(id));
   }
 
-  @ResponseBody
   @GetMapping(value = "save")
-  public double saveMoney(@RequestParam("money") double money, @RequestParam("id") String id) {
-    return accountService.saveMoney(id, money);
+  public ResponseEntity<String> saveMoney(@RequestParam("money") double money, @RequestParam("id") String id) {
+    if (money < 0 || StringUtils.isBlank(id)) {
+      return ResponseEntity.badRequest().build();
+    }
+    return ResponseEntity.ok(accountService.saveMoney(id, money) + "");
   }
 
-  @ResponseBody
   @GetMapping(value = "consume")
-  public double consume(@RequestParam("money") double money, @RequestParam("id") String id) {
-    return accountService.consume(id, money);
+  public ResponseEntity<String> consume(@RequestParam("money") double money, @RequestParam("id") String id) {
+    if (money < 0 || StringUtils.isBlank(id)) {
+      return ResponseEntity.badRequest().build();
+    }
+    double restMoney = accountService.consume(id, money);
+    if (restMoney < 0) {
+      return ResponseEntity.status(HttpStatus.EXPECTATION_FAILED).body(" 余额不足 ");
+    }
+    return ResponseEntity.ok(restMoney + "");
   }
+
 }
